backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
google-generativeai==0.7.2